mempubsub: move to internal and rename (#711)
